Moved object path function to utils folder
